+++ MultiTrain/classification/classification_models.py
@@ -1,5 +1,7 @@
 from collections import Counter
 from operator import __setitem__
+from typing import Union
+
 import seaborn as sns
 import plotly.express as px
 
@@ -75,8 +77,9 @@ from MultiTrain.methods.multitrain_methods import (
     img_plotly,
     kf_best_model,
     write_to_excel,
-    _check_target,
+    _check_target, _get_cat_num, _fill, _fill_columns, _dummy,
 )
+
 from skopt import BayesSearchCV
 from sklearn.model_selection import (
     train_test_split,
@@ -345,6 +348,7 @@ class MultiClassifier:
             method = self.over_under_methods[index_]
             return method
 
+
     def split(
         self,
         X: any,
@@ -357,6 +361,8 @@ class MultiClassifier:
         normalize: any = None,
         columns_to_scale: list = None,
         n_components: int = None,
+        missing_values: dict = None,
+        encode: Union[str, dict] = None
     ):
 
         global the_y
@@ -372,6 +378,7 @@ class MultiClassifier:
         :param sizeOfTest: define size of test data
         :param randomState: define random state
         :param shuffle_data: If set to True, it sets shuffle to True in train_test_split
+        :param missing_values: Dictionary to fill missing values for categorical and numerical columns, e.g {'cat': 'most_frequent', 'num': 'mean'} where the key 'cat' represents categorical column and the corresponding value represents the strategy used to fill the missing value.
 
         Example
         df = pd.read_csv("nameOfFile.csv")
@@ -394,7 +401,28 @@ class MultiClassifier:
 
         else:
             # values for normalize
-            norm = ["StandardScaler", "MinMaxScaler", "RobustScaler"]
+            norm = ['StandardScaler', 'MinMaxScaler', 'RobustScaler']
+
+            if missing_values:
+                if isinstance(missing_values, dict):
+                    if missing_values['cat'] != 'most_frequent':
+                        raise ValueError(
+                            f"Received value '{missing_values['cat']}', you can only use 'most_frequent' for "
+                            f"categorical columns")
+                    elif missing_values['num'] not in ['mean', 'median', 'most_frequent']:
+                        raise ValueError(
+                            f"Received value '{missing_values['num']}', you can only use one of ['mean', 'median', "
+                            f"'most_frequent'] for numerical columns")
+                    categorical_values, numerical_values = _get_cat_num(missing_values)
+                    cat, num = _fill(categorical_values, numerical_values)
+                    X = _fill_columns(cat, num, X)
+
+                else:
+                    raise TypeError(
+                        f'missing_values parameter can only be of type dict, type {type(missing_values)} received')
+
+            X = _dummy(X, encode)
+
             if strat is True:
 
                 if shuffle_data is False:
@@ -1008,7 +1036,7 @@ class MultiClassifier:
                         try:
                             model[i].fit(X_tr_, y_tr_)
                         except ValueError:
-                            logger.error(f"{model[i]} has an issue")
+                            logger.error(f'{model[i]} unable to fit properly')
                             pass
 
                     end = time.time()
@@ -1051,7 +1079,7 @@ class MultiClassifier:
                                 pred_train = pipeline.predict(X_tr)
 
                         except Exception:
-                            logger.error(f"{model[i]} has an issue")
+                            logger.error(f'{model[i]} unable to fit properly')
                             pass
 
                     elif vectorizer == "tfidf":
@@ -1080,7 +1108,7 @@ class MultiClassifier:
                                 pipeline.fit(X_tr, y_tr)
 
                         except Exception:
-                            logger.error(f"{model[i]} cannot fit properly")
+                            logger.error(f'{model[i]} unable to fit properly')
                             pass
 
                     end = time.time()