+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # WHAT'S NEW
 
+# 0.12.0 
+- If model is unable to properly compute metrics, it's value is replaced with np.nan
+- Added parameter in split method to encode categorical columns
 # 0.11.21 - BUG FIX
 - Added missing requirement for kaleido engine
 # 0.11.0 - PATCH RELEASE